# Agrego Logger
# Escribe resultado de programa en log.txt
# Lee de app.config el path del archivo GDF

+++ tp1/src/tp1/tp1/Influences/ShortestPathsExtension.cs
@@ -31,7 +31,7 @@ namespace TP1.Influences
 
             foreach (var node in graph.Nodes)
             {
-                const long infinity = int.MaxValue - 1;
+                const long infinity = long.MaxValue - 1;
                 if (!node.Equals(source))
                 {
                     distanceTo[node.Id] = infinity;
